Drop jobdir/meta file + refactor bartender.filesystem modules to staging and walk_dir module
